#2653 convert eq_ts to datatype to prevent incompatible dtype
Removes warnign from pandas:

C:\Users\kenord\pyc\Energibruksmodell\ebm\energy_consumption.py:133: FutureWarning: Setting an item of incompatible dtype is deprecated and will raise an error in a future version of pandas. Value '[1192689.112636998 1192689.1126369874 1192689.1126369874 ...
 39.059948338457446 27.98687890677554 14.966896054888903]' has dtype incompatible with float64, please explicitly cast to a compatible dtype first.
  df.loc[vann_slice, HEAT_PUMP] = df.loc[vann_slice, ADJUSTED_REQUIREMENT] * df.loc[vann_slice, GRUNNLAST_ANDEL]

+++ ebm/energy_consumption.py
@@ -88,7 +88,7 @@ class EnergyConsumption:
         df = self._merge_energy_requirement_and_heating_systems(energy_requirements)
 
         # Make column eq_ts for tek_share adjusted energy requirement
-        df[ADJUSTED_REQUIREMENT] = df.energy_requirement * df[TEK_SHARES]
+        df[ADJUSTED_REQUIREMENT] = (df.energy_requirement * df[TEK_SHARES]).astype(float)
 
         # Zero fill columns before calculating to prevent NaN from messing up sums
         df.loc[:, [HEATING_RV_GRUNNLAST, HEATING_RV_SPISSLAST, HEATIG_RV_EKSTRALAST, DHW_TV, COOLING_KV, OTHER_SV,
